3. TESTE DE BANCO DE DADOS/test_database_python/requirements.txt
python-dotenv
requests
mysql-connector-python
pandas
